Crate rfc1055

This crate provides a library for encoding and decoding RFC1055 Serial Line Internet Protocol (SLIP) frames. The crate is intended to be used in an embedded environment, so it requires neither std nor alloc, and it makes use of the non-blocking nb crate.

For more information on RFC1055, see https://datatracker.ietf.org/doc/html/rfc1055.

Getting Started

First, read the documentation for Decoder::read, Decoder::new, Encoder::write, and Encoder::new. After that, read the source code for the bundled command line tool and the tests.

Examples

#![macro_use]
use rfc1055::nb;
use rfc1055::nb::block;
use rfc1055::{Decoder,Encoder,END};

fn uart_getchar() -> nb::Result<u8, ()> {
    // replace with your UART implementation
    Ok(END)
}

fn uart_putchar(byte: u8) -> nb::Result<(), ()> {
    // replace with your UART implementation
    Ok(())
}

fn main() {
    let mut decoder = Decoder::new(uart_getchar, true);
    let mut encoder = Encoder::new(uart_putchar);
    let mut buffer: [u8; 1024] = [0; 1024]; // Max frame size of 1 kB

    'top: loop {
        // Get the next frame
        let frame_length = {
            let mut num_read = 0;
            loop {
                num_read += match block!(decoder.read(&mut buffer[num_read..])) {
                    Ok(0) => { break; }, // `read` returns 0 to indicate the end of a frame
                    Ok(n) => n,
                    Err(_) => { continue 'top; },
                };
            }
            num_read
        };

        // Do some stuff with the received frame
        // ...

        // Echo the frame back on the transmitter
        let mut num_written = 0;
        loop {
            num_written += match block!(encoder.write(&buffer[num_written..frame_length])) {
                Ok(n) => n,
                Err(_) => { continue 'top; },
            };

            if num_written == frame_length {
                break;
            }
        }

        break;
    }
}

Command Line Interface

This library also comes with a command line interface that is built when the build-binary feature is enabled. For example:

> cargo build --features build-binary
> cd target/debug
> echo 'hello world!' | ./rfc1055-cli encode | tee /tmp/encoded_data.bin | ./rfc1055-cli decode
hello world!
> xxd /tmp/encoded_data.bin
00000000: 6865 6c6c 6f20 776f 726c 6421 0ac0       hello world!..

Re-exports

pub use nb;

Structs

Decoder

An RFC1055 decoder. It contains the current state of the decoder as well as a pointer to a read routine that fetches individual bytes from some source.

Encoder

An RFC1055 encoder. It contains the state of the encoder as well as a pointer to a write routine that writes individual bytes to some sink.

Enums

DecodeError

Errors that can occur in Decoder::read

EncodeError

Errors that occur in Encoder::write

Constants

END

The u8 value that signals the end of an RFC1055 frame.

ESC

The u8 value that signals the start of an escape sequence.

ESC_END

ESC followed by ESC_END indicates a value of END was sent on the line

ESC_ESC

ESC followed by ESC_ESC indicates a value of ESC was sent on the line

Functions

decode_from_buffer

Create a decoder that reads from a u8 slice.

encode_to_buffer

Create an encoder which writes to a mutable u8 slice.